% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{add_tag}
\alias{add_tag}
\title{Add_tag
Add metadata html span tag (only run in RMarkdown)}
\usage{
add_tag(tagtype, txt)
}
\arguments{
\item{tagtype}{The name of the EML element that will be tagged}

\item{txt}{The txt for the tag}
}
\value{
Output: html span tag
}
\description{
Add_tag
Add metadata html span tag (only run in RMarkdown)
}
