% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_title_tag.R
\name{add_tag_title}
\alias{add_tag_title}
\title{Add_title_tag
Add metadata title html tag (only run in RMarkdown)}
\usage{
add_tag_title(Title)
}
\arguments{
\item{Title}{The title of the project}
}
\value{
Output: html div tag "LN_Title"
}
\description{
Add_title_tag
Add metadata title html tag (only run in RMarkdown)
}
