% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_folder_structure.R
\name{build_folder_structure}
\alias{build_folder_structure}
\title{Build_folder_structure
This function builds the data folder structure in the working directory}
\usage{
build_folder_structure(project_name = project_name)
}
\arguments{
\item{project_name}{The name of the project (it can not contain spaces; use "_" instead)}
}
\value{
A set of folders and a rmarkdown metadata template
}
\description{
Build_folder_structure
This function builds the data folder structure in the working directory
}
