% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_time_data.R
\name{Check_date}
\alias{Check_date}
\title{Check_date
Look for a date column}
\usage{
Check_date(dataframe)
}
\arguments{
\item{dataframe}{dataframe}
}
\value{
Output: list of valid cols (is.Date)
}
\description{
Check_date
Look for a date column
}
