% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_locations.R
\name{check_locations}
\alias{check_locations}
\title{Check_locations
Map out the locations, to see if they}
\usage{
check_locations(lat, lon)
}
\arguments{
\item{X}{lat lon coords cols}
}
\value{
Output: map of locations highlight outliers??
}
\description{
Check_locations
Map out the locations, to see if they
}
