% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_taxonomy.R
\name{check_taxonomy}
\alias{check_taxonomy}
\title{Check_taxonomy
Compare to NBIC artslister}
\usage{
check_taxonomy(X)
}
\arguments{
\item{X}{taxonomy column from dataframe}
}
\value{
Output: list of non-valid names (just to check not stop the workflow)
}
\description{
Check_taxonomy
Compare to NBIC artslister
}
