% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_time_data.R
\name{check_time_data}
\alias{check_time_data}
\title{Check_time_data
compare with standard time_data format, and look for outliers}
\usage{
check_time_data(X)
}
\arguments{
\item{X}{time_data column from dataframe}
}
\value{
Output: list of non-valid rows (just to check not stop the workflow)
}
\description{
Check_time_data
compare with standard time_data format, and look for outliers
}
