% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCMetadata.R
\docType{class}
\name{DwCMetadata}
\alias{DwCMetadata}
\title{R6 class that represents metadata in DwC}
\format{
\code{\link{R6Class}} object.
}
\description{
R6 class that represents metadata in DwC

R6 class that represents metadata in DwC
}
\details{
The \code{DwCMetadata} class exists as a handling class for the encapsulation and
handling of metadata being presented
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-importFromLivingNorwayHTML}{\code{DwCMetadata$importFromLivingNorwayHTML()}}
\item \href{#method-importFromLivingNorwayRMD}{\code{DwCMetadata$importFromLivingNorwayRMD()}}
\item \href{#method-clone}{\code{DwCMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importFromLivingNorwayHTML"></a>}}
\if{latex}{\out{\hypertarget{method-importFromLivingNorwayHTML}{}}}
\subsection{Method \code{importFromLivingNorwayHTML()}}{
Retrieve metadata information from a HTML file that has been created with a
Living Norway HTML tag schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$importFromLivingNorwayHTML(fileLocation, fileEncoding = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileLocation}}{A \code{character} scalar containing the location of the
HTML file}

\item{\code{fileEncoding}}{A character string. If non-empty, declares the encoding to be used on a file so the
character data can be re-encoded as they are written}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importFromLivingNorwayRMD"></a>}}
\if{latex}{\out{\hypertarget{method-importFromLivingNorwayRMD}{}}}
\subsection{Method \code{importFromLivingNorwayRMD()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$importFromLivingNorwayRMD(fileLocation, fileEncoding)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
