% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCTerm.R
\name{DwCTerm}
\alias{DwCTerm}
\title{Initialise a DwCTerm object}
\description{
Initialise a DwCTerm object

Initialise a DwCTerm object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DwCTerm$new()}}
\item \href{#method-getQualifiedName}{\code{DwCTerm$getQualifiedName()}}
\item \href{#method-isDeprecated}{\code{DwCTerm$isDeprecated()}}
\item \href{#method-print}{\code{DwCTerm$print()}}
\item \href{#method-getTermName}{\code{DwCTerm$getTermName()}}
\item \href{#method-getNamespaceName}{\code{DwCTerm$getNamespaceName()}}
\item \href{#method-getTermIRI}{\code{DwCTerm$getTermIRI()}}
\item \href{#method-getTermVersionIRI}{\code{DwCTerm$getTermVersionIRI()}}
\item \href{#method-getDateModified}{\code{DwCTerm$getDateModified()}}
\item \href{#method-getLabel}{\code{DwCTerm$getLabel()}}
\item \href{#method-getIsReplacedBy}{\code{DwCTerm$getIsReplacedBy()}}
\item \href{#method-getDefinition}{\code{DwCTerm$getDefinition()}}
\item \href{#method-getNotes}{\code{DwCTerm$getNotes()}}
\item \href{#method-getType}{\code{DwCTerm$getType()}}
\item \href{#method-getExamples}{\code{DwCTerm$getExamples()}}
\item \href{#method-getTermInformationLN}{\code{DwCTerm$getTermInformationLN()}}
\item \href{#method-getExecCommitteeDecisions}{\code{DwCTerm$getExecCommitteeDecisions()}}
\item \href{#method-getMiscInformation}{\code{DwCTerm$getMiscInformation()}}
\item \href{#method-getTermDef}{\code{DwCTerm$getTermDef()}}
\item \href{#method-clone}{\code{DwCTerm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$new(
  termName,
  namespaceName = character(),
  termIRI = character(),
  termVersionIRI = character(),
  dateModified = character(),
  label = character(),
  isReplacedBy = character(),
  definition = character(),
  notes = character(),
  type = character(),
  examples = character(),
  termInformationLN = character(),
  execCommitteeDecisions = character(),
  miscInformation = character(),
  termDef = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{termName}}{A \code{character} scalar containing the name of the term}

\item{\code{namespaceName}}{A \code{character} scalar containing the name of the namespace containing the term}

\item{\code{termIRI}}{A \code{character} scalar containing the Internationalized Resource Identifier (IRI) of the definition of the term}

\item{\code{termVersionIRI}}{A \code{character} scalar containing the Internationalized Resource Identifier (IRI) of the definition of the current version of the term}

\item{\code{dateModified}}{A \code{character} scalar containing date information of the last modification of the term definition}

\item{\code{label}}{A \code{character} scalar containing a label for the term (used in casual language)}

\item{\code{isReplacedBy}}{A \code{character} scalar containing the Internationalized Resource Identifier (IRI) of the definition of the term that replaced this term (used
if the term is deprecated)}

\item{\code{definition}}{A \code{character} scalar containing a brief description of the purpose of the term}

\item{\code{notes}}{A \code{character} vector containing extra information about the usage of the term}

\item{\code{type}}{A \code{character} scalar containing the type of term defined}

\item{\code{examples}}{A \code{character} vector giving a set of example values for the term}

\item{\code{termInformationLN}}{A \code{characer} vector giving supplementary information for the term provided by the \url{https://livingnorway.no/}{Living Norway} initiative}

\item{\code{execCommitteeDecisions}}{A \code{character} vector of links to decisions made by executive committees on the usage of the term}

\item{\code{miscInformation}}{A \code{character} vector providing supplementary information for the term}

\item{\code{termDef}}{A \code{character} scalar providing information about the location of the body in charge of definition of the term}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new DwCTerm object
Retrieve the qualified name of the term (the term name with the associated namespace)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getQualifiedName"></a>}}
\if{latex}{\out{\hypertarget{method-getQualifiedName}{}}}
\subsection{Method \code{getQualifiedName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getQualifiedName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The qualified name of the term

Return the deprecation status of the term
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isDeprecated"></a>}}
\if{latex}{\out{\hypertarget{method-isDeprecated}{}}}
\subsection{Method \code{isDeprecated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$isDeprecated()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{logical} scalar that is \code{TRUE} if the term is deprecated and
\code{FALSE} otherwise

Print the term information

Retrieve the term name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTermName"></a>}}
\if{latex}{\out{\hypertarget{method-getTermName}{}}}
\subsection{Method \code{getTermName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getTermName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the term name

Retrieve the namespace name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNamespaceName"></a>}}
\if{latex}{\out{\hypertarget{method-getNamespaceName}{}}}
\subsection{Method \code{getNamespaceName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getNamespaceName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the namespace name

Retrieve the term IRI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTermIRI"></a>}}
\if{latex}{\out{\hypertarget{method-getTermIRI}{}}}
\subsection{Method \code{getTermIRI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getTermIRI()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the term IRI

Retrieve the term version IRI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTermVersionIRI"></a>}}
\if{latex}{\out{\hypertarget{method-getTermVersionIRI}{}}}
\subsection{Method \code{getTermVersionIRI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getTermVersionIRI()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the term version IRI

Retrieve the modification date
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDateModified"></a>}}
\if{latex}{\out{\hypertarget{method-getDateModified}{}}}
\subsection{Method \code{getDateModified()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getDateModified()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the modification date

Retrieve the label
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLabel"></a>}}
\if{latex}{\out{\hypertarget{method-getLabel}{}}}
\subsection{Method \code{getLabel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getLabel()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the label

Retrieve the replacement IRI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIsReplacedBy"></a>}}
\if{latex}{\out{\hypertarget{method-getIsReplacedBy}{}}}
\subsection{Method \code{getIsReplacedBy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getIsReplacedBy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the IRI of the replacement term

Retrieve the definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDefinition"></a>}}
\if{latex}{\out{\hypertarget{method-getDefinition}{}}}
\subsection{Method \code{getDefinition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getDefinition()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the term definition

Retrieve the term notes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNotes"></a>}}
\if{latex}{\out{\hypertarget{method-getNotes}{}}}
\subsection{Method \code{getNotes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getNotes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} vector containing the term notes

Retrieve the type of the term
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getType"></a>}}
\if{latex}{\out{\hypertarget{method-getType}{}}}
\subsection{Method \code{getType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the term type
Retrieve examples of the use of the term
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExamples"></a>}}
\if{latex}{\out{\hypertarget{method-getExamples}{}}}
\subsection{Method \code{getExamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getExamples()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} vector containing examples of use of the term

Retrieve supplementary information about the term provided by Living Norway
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTermInformationLN"></a>}}
\if{latex}{\out{\hypertarget{method-getTermInformationLN}{}}}
\subsection{Method \code{getTermInformationLN()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getTermInformationLN()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} vector containing the Living Norway supplementary information

Retrieve executive committee decisions about the use of the term
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExecCommitteeDecisions"></a>}}
\if{latex}{\out{\hypertarget{method-getExecCommitteeDecisions}{}}}
\subsection{Method \code{getExecCommitteeDecisions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getExecCommitteeDecisions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} vector containing the executive committee decisions

Retrieve miscellaneous information on the usage of the term
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMiscInformation"></a>}}
\if{latex}{\out{\hypertarget{method-getMiscInformation}{}}}
\subsection{Method \code{getMiscInformation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getMiscInformation()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} vector containing the miscellaneous information

Retrieve the source of the term definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTermDef"></a>}}
\if{latex}{\out{\hypertarget{method-getTermDef}{}}}
\subsection{Method \code{getTermDef()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$getTermDef()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} scalar containing the source of the term definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCTerm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
