% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\docType{data}
\name{fmt}
\alias{fmt}
\title{add_Tag
Add metadata html tag (only run in RMarkdown)}
\format{
An object of class \code{character} of length 1.
}
\usage{
fmt
}
\arguments{
\item{tagtype}{The name of the EML element that will be tagged}

\item{txt}{The txt for the tag}
}
\value{
Output: html tag
}
\description{
add_Tag
Add metadata html tag (only run in RMarkdown)
}
\keyword{datasets}
