% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFEnvironmentLookup.R
\name{GBIFClassLookup}
\alias{GBIFClassLookup}
\title{Lookup up a GBIF class environment from a qualified name}
\usage{
GBIFClassLookup(qualName)
}
\arguments{
\item{qualName}{A \code{character} of qualifed names for the GBIF class to lookup}
}
\value{
A \code{list} of environments representing \code{R6} classes derived from \code{DwCGeneric}
corresponding to each element in \code{qualName}.
}
\description{
Lookup up a GBIF class environment from a qualified name
}
