% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNHTMLFunctions.R
\name{generateTagID}
\alias{generateTagID}
\title{Check tag ID
Check values given as IDs for EML tag. It the entry is NA then a unique ID is generated}
\usage{
generateTagID(tagID)
}
\arguments{
\item{tagID}{A \code{character} scalar for with the ID code}
}
\value{
A \code{character} scalar containing the processed ID code
}
\description{
Check tag ID
Check values given as IDs for EML tag. It the entry is NA then a unique ID is generated
}
\seealso{
\code{\link[uuid::UUIDgenerate]{UUIDgenerate}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}; Matthew Grainger, \email{matthew.grainger@nina.no}
}
