% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_metadata.R
\name{get_data_metadata}
\alias{get_data_metadata}
\title{get_data_metadata
read the metadata from a rmarkdown file}
\usage{
get_data_metadata(file)
}
\arguments{
\item{X}{an RMarkdown file}
}
\value{
Output: text string of yaml information
}
\description{
get_data_metadata
read the metadata from a rmarkdown file
}
