% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{get_files
Get a list of files in the data folders}
\usage{
get_files(project_name)
}
\arguments{
\item{project_name}{the project name}
}
\value{
Output: temporal extent chart?List of data files, and potentially a description of the field names/types.
}
\description{
get_files
Get a list of files in the data folders
}
