% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Nor_geographic_extent.R,
%   R/get_geographic_extent.R
\name{get_NOR_geographic_extent}
\alias{get_NOR_geographic_extent}
\title{Get the geographic extent of a dataframe with geographic UTM XY coordinates}
\usage{
get_NOR_geographic_extent(df, X, Y, Code)

get_NOR_geographic_extent(df, X, Y, Code)
}
\arguments{
\item{df}{the dataframe}

\item{X}{latitude column indicated by data$X}

\item{Y}{longitude column indicated by data$Y}

\item{Code}{Code to convert UTM to lat lon}
}
\value{
A ggplot map of the coordinates

A ggplot map of the coordinates
}
\description{
This function plots a map of the UTM X Y coordinates from a datafile specified by the user.

This function plots a map of the UTM X Y coordinates from a datafile specified by the user.
}
