% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_title.R
\name{get_project_title}
\alias{get_project_title}
\title{get project title from working directory
This function gets the project title from the working directory for use in the metadata.rmd}
\usage{
get_project_title()
}
\value{
the project title as a charcter vector
}
\description{
get project title from working directory
This function gets the project title from the working directory for use in the metadata.rmd
}
