% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa_cover.R
\name{get_taxa_cover}
\alias{get_taxa_cover}
\title{get_taxa_cover
Get taxonomic coverage from raw data / mapped data.}
\usage{
get_taxa_cover(
  data,
  TaxaField = "SpeciesName",
  TaxaLevel = "Species",
  addFreq = "Yes",
  addPlot = "Yes"
)
}
\arguments{
\item{data}{Dataset including a column with taxonomic information for the record/occurence}

\item{TaxaField}{The field (column) in the data set containing information about the taxonomic group of the record/occurence}

\item{TaxaLevel}{Taxonomic level (e.g. Species, Genus)}

\item{addFreq}{Control argument - include number of records in the table listing the taxonomic coverage}

\item{addPlot}{Control argument - add a barplot with number of records for each taxa defined above}
}
\value{
Output: list of taxa. Potential chart displaying proportion of occurences from different taxa.
}
\description{
get_taxa_cover
Get taxonomic coverage from raw data / mapped data.
}
