% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCMetadata.R
\name{getDefaultEMLDefinitionInfo}
\alias{getDefaultEMLDefinitionInfo}
\title{Retrieve the default namespace definition information used for EML document
initialisation}
\usage{
getDefaultEMLDefinitionInfo()
}
\value{
A \code{character} vector containing named elements that will be added to the
call of the \code{\link[xml2::xml_new_root]{}} function when initialising new
EML documents from Living Norway metadata documents
}
\description{
Retrieve the default namespace definition information used for EML document
initialisation
}
