% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFAmplification.R
\name{getGBIFAmplificationMembers}
\alias{getGBIFAmplificationMembers}
\title{Return a list of properties associated with the GBIF Amplification class}
\usage{
getGBIFAmplificationMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Amplification class
}
\description{
Return a list of properties associated with the GBIF Amplification class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
