% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFAmplification.R
\name{getGBIFAmplificationTerm}
\alias{getGBIFAmplificationTerm}
\title{Amplification term information}
\usage{
getGBIFAmplificationTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Amplification class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
