% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFChronometricAge.R
\name{getGBIFChronometricAgeMembers}
\alias{getGBIFChronometricAgeMembers}
\title{Return a list of properties associated with the GBIF ChronometricAge class}
\usage{
getGBIFChronometricAgeMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF ChronometricAge class
}
\description{
Return a list of properties associated with the GBIF ChronometricAge class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
