% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFChronometricAge.R
\name{getGBIFChronometricAgeTerm}
\alias{getGBIFChronometricAgeTerm}
\title{Return the information of the term associated with the GBIF ChronometricAge class}
\usage{
getGBIFChronometricAgeTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF ChronometricAge class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
