% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFChronometricDate.R
\name{getGBIFChronometricDateMembers}
\alias{getGBIFChronometricDateMembers}
\title{Return a list of properties associated with the GBIF ChronometricDate class}
\usage{
getGBIFChronometricDateMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF ChronometricDate class
}
\description{
Return a list of properties associated with the GBIF ChronometricDate class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
