% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFChronometricDate.R
\name{getGBIFChronometricDateTerm}
\alias{getGBIFChronometricDateTerm}
\title{Return the information of the term associated with the GBIF ChronometricDate class}
\usage{
getGBIFChronometricDateTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF ChronometricDate class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
