% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFCloning.R
\name{getGBIFCloningMembers}
\alias{getGBIFCloningMembers}
\title{Return a list of properties associated with the GBIF Cloning class}
\usage{
getGBIFCloningMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Cloning class
}
\description{
Return a list of properties associated with the GBIF Cloning class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
