% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFCloning.R
\name{getGBIFCloningTerm}
\alias{getGBIFCloningTerm}
\title{Return the information of the term associated with the GBIF Cloning class}
\usage{
getGBIFCloningTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Cloning class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
