% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFDescription.R
\name{getGBIFDescriptionMembers}
\alias{getGBIFDescriptionMembers}
\title{Return a list of properties associated with the GBIF Description class}
\usage{
getGBIFDescriptionMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Description class
}
\description{
Return a list of properties associated with the GBIF Description class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
