% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFDescription.R
\name{getGBIFDescriptionTerm}
\alias{getGBIFDescriptionTerm}
\title{Return the information of the term associated with the GBIF Description class}
\usage{
getGBIFDescriptionTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Description class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
