% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFDistribution.R
\name{getGBIFDistributionMembers}
\alias{getGBIFDistributionMembers}
\title{Return a list of properties associated with the GBIF Distribution class}
\usage{
getGBIFDistributionMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Distribution class
}
\description{
Return a list of properties associated with the GBIF Distribution class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
