% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFDistribution.R
\name{getGBIFDistributionTerm}
\alias{getGBIFDistributionTerm}
\title{Return the information of the term associated with the GBIF Distribution class}
\usage{
getGBIFDistributionTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Distribution class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
