% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFdnaDerivedData.R
\name{getGBIFdnaDerivedDataMembers}
\alias{getGBIFdnaDerivedDataMembers}
\title{Return a list of properties associated with the GBIF dnaDerivedData class}
\usage{
getGBIFdnaDerivedDataMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF dnaDerivedData class
}
\description{
Return a list of properties associated with the GBIF dnaDerivedData class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
