% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFEOLMediaExtension.R
\name{getGBIFEOLMediaExtensionMembers}
\alias{getGBIFEOLMediaExtensionMembers}
\title{Return a list of properties associated with the GBIF EOLMediaExtension class}
\usage{
getGBIFEOLMediaExtensionMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF EOLMediaExtension class
}
\description{
Return a list of properties associated with the GBIF EOLMediaExtension class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
