% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFEOLReferencesExtension.R
\name{getGBIFEOLReferencesExtensionMembers}
\alias{getGBIFEOLReferencesExtensionMembers}
\title{Return a list of properties associated with the GBIF EOLReferencesExtension class}
\usage{
getGBIFEOLReferencesExtensionMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF EOLReferencesExtension class
}
\description{
Return a list of properties associated with the GBIF EOLReferencesExtension class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
