% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFEOLReferencesExtension.R
\name{getGBIFEOLReferencesExtensionTerm}
\alias{getGBIFEOLReferencesExtensionTerm}
\title{Return the information of the term associated with the GBIF EOLReferencesExtension class}
\usage{
getGBIFEOLReferencesExtensionTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF EOLReferencesExtension class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
