% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFEvent.R
\name{getGBIFEventMembers}
\alias{getGBIFEventMembers}
\title{Return a list of properties associated with the GBIF Event class}
\usage{
getGBIFEventMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Event class
}
\description{
Return a list of properties associated with the GBIF Event class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
