% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFEvent.R
\name{getGBIFEventTerm}
\alias{getGBIFEventTerm}
\title{Return the information of the term associated with the GBIF Event class}
\usage{
getGBIFEventTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Event class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
