% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFExtendedMeasurementOrFact.R
\name{getGBIFExtendedMeasurementOrFactMembers}
\alias{getGBIFExtendedMeasurementOrFactMembers}
\title{Return a list of properties associated with the GBIF ExtendedMeasurementOrFact class}
\usage{
getGBIFExtendedMeasurementOrFactMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF ExtendedMeasurementOrFact class
}
\description{
Return a list of properties associated with the GBIF ExtendedMeasurementOrFact class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
