% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFExtendedMeasurementOrFact.R
\name{getGBIFExtendedMeasurementOrFactTerm}
\alias{getGBIFExtendedMeasurementOrFactTerm}
\title{Return the information of the term associated with the GBIF ExtendedMeasurementOrFact class}
\usage{
getGBIFExtendedMeasurementOrFactTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF ExtendedMeasurementOrFact class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
