% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFGelImage.R
\name{getGBIFGelImageMembers}
\alias{getGBIFGelImageMembers}
\title{Return a list of properties associated with the GBIF GelImage class}
\usage{
getGBIFGelImageMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF GelImage class
}
\description{
Return a list of properties associated with the GBIF GelImage class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
