% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFGelImage.R
\name{getGBIFGelImageTerm}
\alias{getGBIFGelImageTerm}
\title{Return the information of the term associated with the GBIF GelImage class}
\usage{
getGBIFGelImageTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF GelImage class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
