% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFGermplasmAccession.R
\name{getGBIFGermplasmAccessionMembers}
\alias{getGBIFGermplasmAccessionMembers}
\title{Return a list of properties associated with the GBIF GermplasmAccession class}
\usage{
getGBIFGermplasmAccessionMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF GermplasmAccession class
}
\description{
Return a list of properties associated with the GBIF GermplasmAccession class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
