% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFGermplasmAccession.R
\name{getGBIFGermplasmAccessionTerm}
\alias{getGBIFGermplasmAccessionTerm}
\title{Return the information of the term associated with the GBIF GermplasmAccession class}
\usage{
getGBIFGermplasmAccessionTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF GermplasmAccession class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
