% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFGermplasmSample.R
\name{getGBIFGermplasmSampleMembers}
\alias{getGBIFGermplasmSampleMembers}
\title{Return a list of properties associated with the GBIF GermplasmSample class}
\usage{
getGBIFGermplasmSampleMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF GermplasmSample class
}
\description{
Return a list of properties associated with the GBIF GermplasmSample class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
