% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFIdentification.R
\name{getGBIFIdentificationMembers}
\alias{getGBIFIdentificationMembers}
\title{Return a list of properties associated with the GBIF Identification class}
\usage{
getGBIFIdentificationMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Identification class
}
\description{
Return a list of properties associated with the GBIF Identification class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
