% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFIdentification.R
\name{getGBIFIdentificationTerm}
\alias{getGBIFIdentificationTerm}
\title{Return the information of the term associated with the GBIF Identification class}
\usage{
getGBIFIdentificationTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Identification class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
