% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFIdentifier.R
\name{getGBIFIdentifierMembers}
\alias{getGBIFIdentifierMembers}
\title{Return a list of properties associated with the GBIF Identifier class}
\usage{
getGBIFIdentifierMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Identifier class
}
\description{
Return a list of properties associated with the GBIF Identifier class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
