% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFIdentifier.R
\name{getGBIFIdentifierTerm}
\alias{getGBIFIdentifierTerm}
\title{Return the information of the term associated with the GBIF Identifier class}
\usage{
getGBIFIdentifierTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Identifier class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
