% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFImage.R
\name{getGBIFImageMembers}
\alias{getGBIFImageMembers}
\title{Return a list of properties associated with the GBIF Image class}
\usage{
getGBIFImageMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Image class
}
\description{
Return a list of properties associated with the GBIF Image class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
