% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFImage.R
\name{getGBIFImageTerm}
\alias{getGBIFImageTerm}
\title{Return the information of the term associated with the GBIF Image class}
\usage{
getGBIFImageTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Image class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
