% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFLoan.R
\name{getGBIFLoanMembers}
\alias{getGBIFLoanMembers}
\title{Return a list of properties associated with the GBIF Loan class}
\usage{
getGBIFLoanMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Loan class
}
\description{
Return a list of properties associated with the GBIF Loan class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
