% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFLoan.R
\name{getGBIFLoanTerm}
\alias{getGBIFLoanTerm}
\title{Return the information of the term associated with the GBIF Loan class}
\usage{
getGBIFLoanTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Loan class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
