% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMaterialSample.R
\name{getGBIFMaterialSampleMembers}
\alias{getGBIFMaterialSampleMembers}
\title{Return a list of properties associated with the GBIF MaterialSample class}
\usage{
getGBIFMaterialSampleMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF MaterialSample class
}
\description{
Return a list of properties associated with the GBIF MaterialSample class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
