% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMeasurementOrFact.R
\name{getGBIFMeasurementOrFactMembers}
\alias{getGBIFMeasurementOrFactMembers}
\title{Return a list of properties associated with the GBIF MeasurementOrFact class}
\usage{
getGBIFMeasurementOrFactMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF MeasurementOrFact class
}
\description{
Return a list of properties associated with the GBIF MeasurementOrFact class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
