% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMeasurementOrFact.R
\name{getGBIFMeasurementOrFactTerm}
\alias{getGBIFMeasurementOrFactTerm}
\title{Return the information of the term associated with the GBIF MeasurementOrFact class}
\usage{
getGBIFMeasurementOrFactTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF MeasurementOrFact class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
