% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMeasurementScore.R
\name{getGBIFMeasurementScoreMembers}
\alias{getGBIFMeasurementScoreMembers}
\title{Return a list of properties associated with the GBIF MeasurementScore class}
\usage{
getGBIFMeasurementScoreMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF MeasurementScore class
}
\description{
Return a list of properties associated with the GBIF MeasurementScore class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
