% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMeasurementScore.R
\name{getGBIFMeasurementScoreTerm}
\alias{getGBIFMeasurementScoreTerm}
\title{Return the information of the term associated with the GBIF MeasurementScore class}
\usage{
getGBIFMeasurementScoreTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF MeasurementScore class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
