% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMeasurementTrait.R
\name{getGBIFMeasurementTraitMembers}
\alias{getGBIFMeasurementTraitMembers}
\title{Return a list of properties associated with the GBIF MeasurementTrait class}
\usage{
getGBIFMeasurementTraitMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF MeasurementTrait class
}
\description{
Return a list of properties associated with the GBIF MeasurementTrait class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
