% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMeasurementTrait.R
\name{getGBIFMeasurementTraitTerm}
\alias{getGBIFMeasurementTraitTerm}
\title{Return the information of the term associated with the GBIF MeasurementTrait class}
\usage{
getGBIFMeasurementTraitTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF MeasurementTrait class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
