% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMeasurementTrial.R
\name{getGBIFMeasurementTrialMembers}
\alias{getGBIFMeasurementTrialMembers}
\title{Return a list of properties associated with the GBIF MeasurementTrial class}
\usage{
getGBIFMeasurementTrialMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF MeasurementTrial class
}
\description{
Return a list of properties associated with the GBIF MeasurementTrial class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
