% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMeasurementTrial.R
\name{getGBIFMeasurementTrialTerm}
\alias{getGBIFMeasurementTrialTerm}
\title{Return the information of the term associated with the GBIF MeasurementTrial class}
\usage{
getGBIFMeasurementTrialTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF MeasurementTrial class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
