% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMultimedia.R
\name{getGBIFMultimediaMembers}
\alias{getGBIFMultimediaMembers}
\title{Return a list of properties associated with the GBIF Multimedia class}
\usage{
getGBIFMultimediaMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Multimedia class
}
\description{
Return a list of properties associated with the GBIF Multimedia class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
