% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFMultimedia.R
\name{getGBIFMultimediaTerm}
\alias{getGBIFMultimediaTerm}
\title{Return the information of the term associated with the GBIF Multimedia class}
\usage{
getGBIFMultimediaTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Multimedia class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
