% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFOccurrence.R
\name{getGBIFOccurrenceMembers}
\alias{getGBIFOccurrenceMembers}
\title{Return a list of properties associated with the GBIF Occurrence class}
\usage{
getGBIFOccurrenceMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Occurrence class
}
\description{
Return a list of properties associated with the GBIF Occurrence class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
