% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFPermit.R
\name{getGBIFPermitMembers}
\alias{getGBIFPermitMembers}
\title{Return a list of properties associated with the GBIF Permit class}
\usage{
getGBIFPermitMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Permit class
}
\description{
Return a list of properties associated with the GBIF Permit class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
