% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFPreparation.R
\name{getGBIFPreparationMembers}
\alias{getGBIFPreparationMembers}
\title{Return a list of properties associated with the GBIF Preparation class}
\usage{
getGBIFPreparationMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Preparation class
}
\description{
Return a list of properties associated with the GBIF Preparation class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
