% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFPreparation.R
\name{getGBIFPreparationTerm}
\alias{getGBIFPreparationTerm}
\title{Return the information of the term associated with the GBIF Preparation class}
\usage{
getGBIFPreparationTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Preparation class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
