% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFPreservation.R
\name{getGBIFPreservationMembers}
\alias{getGBIFPreservationMembers}
\title{Return a list of properties associated with the GBIF Preservation class}
\usage{
getGBIFPreservationMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Preservation class
}
\description{
Return a list of properties associated with the GBIF Preservation class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
