% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFReference.R
\name{getGBIFReferenceMembers}
\alias{getGBIFReferenceMembers}
\title{Return a list of properties associated with the GBIF Reference class}
\usage{
getGBIFReferenceMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Reference class
}
\description{
Return a list of properties associated with the GBIF Reference class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
