% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFReference.R
\name{getGBIFReferenceTerm}
\alias{getGBIFReferenceTerm}
\title{Return the information of the term associated with the GBIF Reference class}
\usage{
getGBIFReferenceTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Reference class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
