% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFReleve.R
\name{getGBIFReleveMembers}
\alias{getGBIFReleveMembers}
\title{Return a list of properties associated with the GBIF Releve class}
\usage{
getGBIFReleveMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Releve class
}
\description{
Return a list of properties associated with the GBIF Releve class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
