% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFReleve.R
\name{getGBIFReleveTerm}
\alias{getGBIFReleveTerm}
\title{Return the information of the term associated with the GBIF Releve class}
\usage{
getGBIFReleveTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Releve class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
