% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFResourceRelationship.R
\name{getGBIFResourceRelationshipMembers}
\alias{getGBIFResourceRelationshipMembers}
\title{Return a list of properties associated with the GBIF ResourceRelationship class}
\usage{
getGBIFResourceRelationshipMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF ResourceRelationship class
}
\description{
Return a list of properties associated with the GBIF ResourceRelationship class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
