% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFResourceRelationship.R
\name{getGBIFResourceRelationshipTerm}
\alias{getGBIFResourceRelationshipTerm}
\title{Return the information of the term associated with the GBIF ResourceRelationship class}
\usage{
getGBIFResourceRelationshipTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF ResourceRelationship class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
