% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFSpeciesProfile.R
\name{getGBIFSpeciesProfileMembers}
\alias{getGBIFSpeciesProfileMembers}
\title{Return a list of properties associated with the GBIF SpeciesProfile class}
\usage{
getGBIFSpeciesProfileMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF SpeciesProfile class
}
\description{
Return a list of properties associated with the GBIF SpeciesProfile class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
