% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFSpeciesProfile.R
\name{getGBIFSpeciesProfileTerm}
\alias{getGBIFSpeciesProfileTerm}
\title{Return the information of the term associated with the GBIF SpeciesProfile class}
\usage{
getGBIFSpeciesProfileTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF SpeciesProfile class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
