% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFTaxon.R
\name{getGBIFTaxonMembers}
\alias{getGBIFTaxonMembers}
\title{Return a list of properties associated with the GBIF Taxon class}
\usage{
getGBIFTaxonMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF Taxon class
}
\description{
Return a list of properties associated with the GBIF Taxon class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
