% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFTaxon.R
\name{getGBIFTaxonTerm}
\alias{getGBIFTaxonTerm}
\title{Return the information of the term associated with the GBIF Taxon class}
\usage{
getGBIFTaxonTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF Taxon class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
