% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFTypesAndSpecimen.R
\name{getGBIFTypesAndSpecimenMembers}
\alias{getGBIFTypesAndSpecimenMembers}
\title{Types and specimen member information}
\usage{
getGBIFTypesAndSpecimenMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF TypesAndSpecimen class
}
\description{
Return a list of properties associated with the GBIF TypesAndSpecimen class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
