% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFTypesAndSpecimen.R
\name{getGBIFTypesAndSpecimenTerm}
\alias{getGBIFTypesAndSpecimenTerm}
\title{Return the information of the term associated with the GBIF TypesAndSpecimen class}
\usage{
getGBIFTypesAndSpecimenTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF TypesAndSpecimen class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
