% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFVernacularName.R
\name{getGBIFVernacularNameMembers}
\alias{getGBIFVernacularNameMembers}
\title{Vernacular name member information}
\usage{
getGBIFVernacularNameMembers()
}
\value{
A \code{list} of \code{DwCTerm} objects, one for each member of the the
GBIF VernacularName class
}
\description{
Return a list of properties associated with the GBIF VernacularName class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
