% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLNportalData.R
\name{getLNportalData}
\alias{getLNportalData}
\title{Function for getting data from the Livng Norway Portal}
\usage{
getLNportalData(datasetID)
}
\arguments{
\item{datasetID}{the dataset unique identifying key}
}
\description{
using the dataset key for a Living Norway dataset (and other GBIF datasets) you can download and extract the elements of the Darwin Core Archive file
}
\examples{
Archive=getLNportalData(datasetID = "b848f1f3-3955-4725-8ad8-e711e4a9e0ac" )
core<-Archive$getCoreTable()
extension<-Archive$getExtensionTables()
}
