% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{getTagGenerationFunctions}
\alias{getTagGenerationFunctions}
\title{getTagGenerationFunctions
Get full list of available LNtags}
\usage{
getTagGenerationFunctions()
}
\value{
list of available LivingNorwayR metadata tags
}
\description{
getTagGenerationFunctions
Get full list of available LNtags
}
