% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCArchive.R
\name{initializeDwCArchive}
\alias{initializeDwCArchive}
\title{Create a new \code{DwCAchive} object}
\usage{
initializeDwCArchive(coreDwC, extDwC = NULL)
}
\arguments{
\item{coreDwC}{Either a \code{DwCGeneric} (or derived class) object that represents the
table that corresponds to the 'core' table.  Alternatively, this parameter can be
\code{character} scalar giving the location of the Darwin core archive file to
initialize the object from}

\item{extDwC}{A \code{list} of \code{DwCGeneric} (or derived class) objects that represent
the tables used as extension objects in the Darwin Core archive.  If \code{coreDwC}
is a character scalar then \code{extDwC} can also be a character scalar that contains the
default file encodings for the files in the Darwin core archive}
}
\value{
A new \code{DwCArchive} object
}
\description{
Create a new \code{DwCAchive} object
}
\seealso{
\code{\link[DwCGeneric]{DwCGeneric}}
}
