% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCGeneric.R
\name{initializeDwCGeneric}
\alias{initializeDwCGeneric}
\title{Initialize a new DwCGeneric object}
\usage{
initializeDwCGeneric(
  classTermInfo,
  associatedTerms,
  objectData,
  idColumnInfo,
  nameAutoMap = FALSE,
  ...
)
}
\arguments{
\item{classTermInfo}{A \code{DwCTerm} object containing the term information for the class}

\item{associatedTerms}{A \code{list} of \code{DwCTerm} objects that contain all the terms associated with the class}

\item{objectData}{A \code{data.frame} containing the data to import into the object}

\item{idColumnInfo}{Either a \code{character} scalar containing the column name of
\code{objectData} or an \code{integer} scalar giving the index of the column of
\code{objectData} that corresponds to the ID variable.  Alternatively, this parameter
may be the qualified name of the Darwin core term for which the appropriately mapped column
will be used as the ID variable}

\item{nameAutoMap}{A \code{logical} scalar that if \code{TRUE} maps the columns of \code{objectData}
to their respective Darwin core terms based on the column names}

\item{...}{A named set of paramaeters corresponding to Darwin core terms associated with the DwCGeneric
class type.  Each is either a \code{character} scalar containing the column name of \code{objectData}
or an \code{integer} scalar giving the index of the column of \code{objectData} that corresponds to the
term.}
}
\value{
A new \code{DwCGeneric} object
}
\description{
Initialize a new DwCGeneric object
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
