% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCMetadata.R
\name{initializeDwCMetadata}
\alias{initializeDwCMetadata}
\title{Initialise a metadata object from an import file}
\usage{
initializeDwCMetadata(fileLocation, fileEncoding = "", fileType = NA)
}
\arguments{
\item{fileLocation}{A \code{character} scalar containing the location of the import file}

\item{fileEncoding}{A \code{character} string. If non-empty, declares the encoding to be used on a file so the
character data can be re-encoded as they are written}

\item{fileType}{A \code{character} scalar stating the type of the file.  File type can be either
\code{"rmarkdown"}, \code{"html"}, \code{"eml"}, or \code{"darwincore"}. If \code{NA} then the
file type will be determined from the file extension}
}
\value{
A new \code{DwCMetadata} object
}
\description{
Initialise a metadata object from an import file
}
\seealso{
\code{\link[DwCMetadata]{DwCMetadata}}
}
