% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCGeneric.R
\name{isDwCGeneric}
\alias{isDwCGeneric}
\title{Test whether an object is derived from DwCGeneric}
\usage{
isDwCGeneric(inOb)
}
\arguments{
\item{inOb}{An object that you want to test}
}
\value{
A \code{logical} scalar that is \code{TRUE} if \code{inOb} is derived from DwCGeneric
}
\description{
Test whether an object is derived from DwCGeneric
}
