% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LN_alternateIdentifier}
\alias{LN_alternateIdentifier}
\title{LN_alternativeIdentifier}
\usage{
LN_alternateIdentifier(alternateIdentifier)
}
\value{
Output: html tag
}
\description{
LN_alternativeIdentifier
}
