% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LN_boundingCoordinates}
\alias{LN_boundingCoordinates}
\title{boundingCoordinates}
\usage{
LN_boundingCoordinates(
  westBoundingCoordinate,
  eastBoundingCoordinate,
  northBoundingCoordinate,
  southBoundingCoordinate,
  hidden = FALSE
)
}
\value{
Output: html tag
}
\description{
boundingCoordinates
}
