% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LN_coverage}
\alias{LN_coverage}
\title{coverage}
\usage{
LN_coverage(
  geographicCoverage,
  temporalCoverage,
  taxonomicCoverage,
  hidden = FALSE
)
}
\value{
Output: html tag
}
\description{
coverage
}
