% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LN_creator}
\alias{LN_creator}
\title{LN_creator
Add creator metadata html tag (only run in RMarkdown)}
\usage{
LN_creator(
  individualName,
  organizationName = NULL,
  positionName = NULL,
  deliveryPoint = NULL,
  city = NULL,
  postalCode = NULL,
  electonicMail = NULL,
  order = NULL
)
}
\value{
Output: html tag
}
\description{
LN_creator
Add creator metadata html tag (only run in RMarkdown)
}
