% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LN_geographicCoverage}
\alias{LN_geographicCoverage}
\title{geographicCoverage}
\usage{
LN_geographicCoverage(geographicDescription, boundingCoordinates)
}
\value{
Output: html tag
}
\description{
geographicCoverage
}
