% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LN_individualName}
\alias{LN_individualName}
\title{LN_individualName
Add individualName metadata html tag (only run in RMarkdown)}
\usage{
LN_individualName(firstName, secondName, hidden = FALSE)
}
\arguments{
\item{firstName}{The givenName of the individual}

\item{secondName}{The surName of the individual}
}
\value{
Output: html tag
}
\description{
LN_individualName
Add individualName metadata html tag (only run in RMarkdown)
}
