% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LN_keywordSet}
\alias{LN_keywordSet}
\title{keywordSet}
\usage{
LN_keywordSet(keyword, keywordThesaurus = NULL, hidden = FALSE)

LN_keywordSet(keyword, keywordThesaurus = NULL, hidden = FALSE)
}
\value{
Output: html tag

Output: html tag
}
\description{
keywordSet

keywordSet
}
