% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LN_metadataProvider}
\alias{LN_metadataProvider}
\title{metadataProvider}
\usage{
LN_metadataProvider(
  individualName,
  organizationName = NULL,
  positionName = NULL,
  deliveryPoint = NULL,
  city = NULL,
  postalCode = NULL,
  electonicMail = NULL,
  userId = NULL,
  hidden = FALSE
)
}
\value{
Output: html tag
}
\description{
metadataProvider
}
