% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LN_taxonomicCoverage}
\alias{LN_taxonomicCoverage}
\title{TaxonomicCoverage}
\usage{
LN_taxonomicCoverage(taxonRankName, taxonRankValue, commonName, hidden = FALSE)
}
\value{
Output: html tag
}
\description{
TaxonomicCoverage
}
