% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LNaddTag}
\alias{LNaddTag}
\title{LNaddTag
Add metadata html tag (only run in RMarkdown)}
\usage{
LNaddTag(tagText, tagType, tagID = NA, parentID = NA, isHidden = FALSE)
}
\arguments{
\item{tagID}{Any unique ID of the tag}

\item{parentID}{Any unique ID of the parent of the tag}

\item{isHidden}{Hide the tag in the rendered html text}

\item{tagTxt}{The text for the tag}

\item{tagtype}{The name of the EML element that will be tagged}
}
\value{
Output: html tag
}
\description{
LNaddTag
Add metadata html tag (only run in RMarkdown)
}
