% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNTagFunctions.R
\name{LNaddTranslation}
\alias{LNaddTranslation}
\title{Add a translation to a data tag}
\usage{
LNaddTranslation(
  tagText = NA,
  lang = "en-US",
  parentID = NA,
  tagID = NA,
  isHidden = TRUE
)
}
\arguments{
\item{tagText}{A \code{character} scalar containing the translated text to enclose in the tag}

\item{lang}{A \code{character} scalar specifying the language attribute to mark the translation text as.  See
\url{https://www.w3.org/International/articles/language-tags/}{the W3C guide to language tags} for the
best format to use for this attribute}

\item{parentID}{A \code{character} scalar containing the ID for the parent tag (if there is one).
\code{NA} denotes that there is no parent for the tag}

\item{tagID}{A \code{character} scalar containing a unique identifier for the tag element. If
the parameter is \code{NA} then a UUID will be generated}

\item{isHidden}{A \code{logical} scalar that, if \code{TRUE}, denotes that the text should be
hidden when rendered}
}
\value{
A \code{character} scalar containing the rendered output including any HTML tags if a
HTML document is being knitted
}
\description{
Add a translation to a data tag
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}; Matthew Grainger, \email{matthew.grainger@nina.no}
}
