% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LNcity}
\alias{LNcity}
\title{LNcity
Add city metadata html tag (only run in RMarkdown)}
\usage{
LNcity(tagText, tagID = NA, parentID = NA, isHidden = FALSE)
}
\arguments{
\item{tagText}{the text of the tag}

\item{tagID}{Any unique ID of the tag}

\item{parentID}{Any unique ID of the parent of the tag}

\item{isHidden}{Hide the tag in the rendered html text}
}
\value{
Output: html tag
}
\description{
LNcity
Add city metadata html tag (only run in RMarkdown)
}
