% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LNdeliveryPoint}
\alias{LNdeliveryPoint}
\title{LNdeliveryPoint
Add deliveryPoint metadata html tag (only run in RMarkdown)}
\usage{
LNdeliveryPoint(tagText, tagID = NA, parentID = NA, isHidden = FALSE)
}
\arguments{
\item{tagText}{the text of the tag}

\item{tagID}{Any unique ID of the tag}

\item{parentID}{Any unique ID of the parent of the tag}

\item{isHidden}{Hide the tag in the rendered html text}
}
\value{
Output: html tag
}
\description{
LNdeliveryPoint
Add deliveryPoint metadata html tag (only run in RMarkdown)
}
