% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{LNtaxonomicClassification}
\alias{LNtaxonomicClassification}
\title{LNtaxonomicClassification
Add taxonomicClassification metadata html tag (only run in RMarkdown)}
\usage{
LNtaxonomicClassification(tagText, tagID = NA, parentID = NA, isHidden = FALSE)
}
\arguments{
\item{tagText}{the text of the tag}

\item{tagID}{Any unique ID of the tag}

\item{parentID}{Any unique ID of the parent of the tag}

\item{isHidden}{Hide the tag in the rendered html text}
}
\value{
Output: html tag
}
\description{
LNtaxonomicClassification
Add taxonomicClassification metadata html tag (only run in RMarkdown)
}
