% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_folders.R
\name{map_files}
\alias{map_files}
\title{map_files
Get a list of files in the data folders}
\usage{
map_files(project_name, addPlot = "Yes")
}
\arguments{
\item{project_name}{the project name}
}
\value{
Output: List of data files, and potentially a description of the field names/types.
}
\description{
map_files
Get a list of files in the data folders
}
