% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_folders.R
\name{map_folders}
\alias{map_folders}
\title{map_folders
Get a list of files in the data folders}
\usage{
map_folders(project_name)
}
\arguments{
\item{project_name}{the project name}
}
\value{
Output: List of data files, and potentially a description of the field names/types.
}
\description{
map_folders
Get a list of files in the data folders
}
