% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNTagFunctions.R
\name{processChildArgs}
\alias{processChildArgs}
\title{Process the child-parent metadata tags
Create a series of child metadata tags based on a naming convention of \code{...} arguments}
\usage{
processChildArgs(parentID, sep = sep, isHiddenDefault = FALSE, ...)
}
\arguments{
\item{parentID}{A \code{character} scalar containing the unique ID of the parent of the tag
(if there is a parent)}

\item{sep}{A \code{character} scalar providing the delimiter to use between the child tags}

\item{isHiddenDefault}{A \code{logical} scalar that set the default \code{isHidden} parameter
for each of the child tags}

\item{...}{A series of parameters defining the child tags to generate and the parameters of the
generation functions}
}
\value{
A \code{character} scalar containing the rendered output including any HTML tags if a
HTML document is being knitted
}
\description{
Process the child-parent metadata tags
Create a series of child metadata tags based on a naming convention of \code{...} arguments
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}; Matthew Grainger, \email{matthew.grainger@nina.no}
}
\keyword{internal}
