% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{read_metadata
read the metadata from a rmarkdown file}
\usage{
read_metadata(file)
}
\arguments{
\item{X}{an RMarkdown file}
}
\value{
Output: text string of yaml information
}
\description{
read_metadata
read the metadata from a rmarkdown file
}
\keyword{internal}
