% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCTerm.R
\name{retrieveDwCClassSpecifications}
\alias{retrieveDwCClassSpecifications}
\title{Retrieve classes and their related term used by Darwin core}
\usage{
retrieveDwCClassSpecifications(
  includeExtensions = TRUE,
  includeDeprecated = FALSE
)
}
\arguments{
\item{includeExtensions}{A \code{logical} scalar that, if \code{TRUE}, instructs the fuction to also
include classes used in registered Darwin core extensions. If \code{FALSE}, only the classes specified by
the Darwin core standard is included}

\item{includeDeprecated}{A \code{logical} scalar that, if \code{TRUE}, instructs the function to also
include terms in the Dawin Core standard that are depracated}
}
\value{
A \code{list} containing one element per class. Each element is itself a \code{list} with the
following named elements:
\itemize{
\item{termInfo}{A \code{DwCTerm} object containing the information of the class term}
\item{compositeTerms}{A \code{list} of \code{DwCTerm} objects for each term that is associated with
the class}
}
}
\description{
Retrieve classes and their related term used by Darwin core
}
\seealso{
\code{\link[retrieveDwCTermSpecifications]{retrieveDwCTermSpecifications}}
\code{\link[DwCTerm]{DwCTerm}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
