% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCTerm.R
\name{retrieveDwCTermSpecifications}
\alias{retrieveDwCTermSpecifications}
\title{Retrieve terms used by Darwin core}
\usage{
retrieveDwCTermSpecifications(
  includeExtensions = TRUE,
  includeDeprecated = FALSE
)
}
\arguments{
\item{includeExtensions}{A \code{logical} scalar that, if \code{TRUE}, instructs the fuction to also
include terms used in registered GBIF extensions to Darwin Core. If \code{FALSE}, only the terms specified by
the Darwin core standard is included}

\item{includeDeprecated}{A \code{logical} scalar that, if \code{TRUE}, instructs the function to also
include terms in the Dawin Core standard that are deprecated}
}
\value{
A \code{list} of \code{DwCTerm} objects representing the terms defined by the Darwin core
standard
}
\description{
Retrieve terms used by Darwin core
}
\seealso{
\code{\link[retrieveDwCClassSpecifications]{retrieveDwCClassSpecifications}}
\code{\link[DwCTerm]{DwCTerm}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
