% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNHTMLFunctions.R
\name{retrieveEMLSchema}
\alias{retrieveEMLSchema}
\title{Retrieve an EML schema and convert it into an R list that can be used for
creation of metadata functions}
\usage{
retrieveEMLSchema(schema)
}
\arguments{
\item{schema}{A \code{character} scalar with the location of the EML schema to import}
}
\value{
A \code{list} containing the schema elements
}
\description{
Retrieve an EML schema and convert it into an R list that can be used for
creation of metadata functions
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}; Matthew Grainger, \email{matthew.grainger@nina.no}
}
