% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCTerm.R
\name{retrieveGBIFClassSpecifications}
\alias{retrieveGBIFClassSpecifications}
\title{Retrieve classes and their related terms used by GBIF}
\usage{
retrieveGBIFClassSpecifications(classOption = "all", includeDeprecated = FALSE)
}
\arguments{
\item{classOption}{A \code{character} scalar that if set to \code{"core"} returns only the classes and the associated terms of GBIF's accepted core types.
If set to \code{"extension"} returns only the classes and associated terms of GBIF's \url{https://tools.gbif.org/dwca-validator/extensions.do}{registered extensions}.
\code{"all"} (the default) returns all of GBIF's registered class types.}

\item{includeDeprecated}{A \code{logical} scalar that, if \code{TRUE}, instructs the function to also include terms in the GBIF classes that are depracated}
}
\value{
A \code{list} containing one element per class. Each element is itself a \code{list} with the
following named elements:
\itemize{
\item{termInfo}{A \code{DwCTerm} object containing the information of the class term}
\item{compositeTerms}{A \code{list} of \code{DwCTerm} objects for each term that is associated with
the class}
}
}
\description{
Retrieve classes and their related terms used by GBIF
}
\seealso{
\code{\link[retrieveDwCTermSpecifications]{retrieveDwCTermSpecifications}}
\code{\link[DwCTerm]{DwCTerm}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
