% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchGBIFMembersAndClasses.R
\name{searchGBIFCoreClasses}
\alias{searchGBIFCoreClasses}
\title{function to search GBIFCoreClasses}
\usage{
searchGBIFCoreClasses(term)
}
\arguments{
\item{term}{the search term}
}
\value{
A subset of the list from the LivingNorwayR::getGBIFCoreClasses() function
}
\description{
function to search GBIFCoreClasses
}
\examples{
searchGBIFCoreClasses("Event")
}
\seealso{
\code{\link[GBIFEvent]{GBIFEvent}}
\code{\link[GBIFOccurrence]{GBIFOccurrence}}
\code{\link[GBIFTaxon]{GBIFTaxon}}
}
