% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchGBIFMembersAndClasses.R
\name{searchGBIFEventMembers}
\alias{searchGBIFEventMembers}
\title{function to search GBIFEventMembers}
\usage{
searchGBIFEventMembers(term)
}
\arguments{
\item{term}{the search term}
}
\value{
A subset of the list from the LivingNorwayR::getGBIFEventMembers() function
}
\description{
function to search GBIFEventMembers
}
\examples{
searchGBIFEventMembers("georef")
}
\seealso{
\code{\link[GBIFEvent]{GBIFEvent}}
}
