% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchGBIFMembersAndClasses.R
\name{searchGBIFExtensionClasses}
\alias{searchGBIFExtensionClasses}
\title{function to search GBIFExtensionClasses}
\usage{
searchGBIFExtensionClasses(term)
}
\arguments{
\item{term}{the search term}
}
\value{
A subset of the list from the LivingNorwayR::getGBIFExtensionClasses() function
}
\description{
function to search GBIFExtensionClasses
}
\examples{
searchGBIFExtensionClasses("Multimedia")
}
\seealso{
\code{\link[GBIFExtendedMeasurementOrFact]{GBIFExtendedMeasurementOrFact}}
\code{\link[GBIFMeasurementOrFact]{GBIFMeasurementOrFact}}
}
