% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchGBIFMembersAndClasses.R
\name{searchGBIFOccurenceMembers}
\alias{searchGBIFOccurenceMembers}
\title{function to search GBIFOccurrenceMembers}
\usage{
searchGBIFOccurenceMembers(term)
}
\arguments{
\item{term}{the search term}
}
\value{
A subset of the list from the LivingNorwayR::getGBIFOccurrenceMembers() function
}
\description{
function to search GBIFOccurrenceMembers
}
\examples{
searchGBIFOccurrenceMembers("tax")
}
\seealso{
\code{\link[GBIFOccurrence]{GBIFOccurrence}}
}
