% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_tag.R
\name{add_Tag}
\alias{add_Tag}
\title{add_Tag
Add metadata html tag (only run in RMarkdown)}
\usage{
add_Tag(tag, txt)
}
\arguments{
\item{txt}{The txt for the tag}

\item{tagtype}{The name of the EML element that will be tagged}
}
\value{
Output: html tag
}
\description{
add_Tag
Add metadata html tag (only run in RMarkdown)
}
