% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCMetadata.R
\docType{class}
\name{DwCMetadata}
\alias{DwCMetadata}
\title{R6 metadata class}
\format{
\code{\link{R6Class}} object.
}
\description{
R6 class that represents metadata in DwC
}
\details{
The \code{DwCMetadata} class exists as a handling class for the encapsulation and
handling of metadata being presented
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DwCMetadata-importFromLivingNorwayHTML}{\code{DwCMetadata$importFromLivingNorwayHTML()}}
\item \href{#method-DwCMetadata-importFromLivingNorwayRMD}{\code{DwCMetadata$importFromLivingNorwayRMD()}}
\item \href{#method-DwCMetadata-importFromEML}{\code{DwCMetadata$importFromEML()}}
\item \href{#method-DwCMetadata-importFromDwCArchive}{\code{DwCMetadata$importFromDwCArchive()}}
\item \href{#method-DwCMetadata-exportToEML}{\code{DwCMetadata$exportToEML()}}
\item \href{#method-DwCMetadata-new}{\code{DwCMetadata$new()}}
\item \href{#method-DwCMetadata-print}{\code{DwCMetadata$print()}}
\item \href{#method-DwCMetadata-getTitle}{\code{DwCMetadata$getTitle()}}
\item \href{#method-DwCMetadata-getCreatorInfo}{\code{DwCMetadata$getCreatorInfo()}}
\item \href{#method-DwCMetadata-getAbstract}{\code{DwCMetadata$getAbstract()}}
\item \href{#method-DwCMetadata-clone}{\code{DwCMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-importFromLivingNorwayHTML"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-importFromLivingNorwayHTML}{}}}
\subsection{Method \code{importFromLivingNorwayHTML()}}{
Retrieve metadata information from a HTML file that has been created with a
Living Norway HTML tag schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$importFromLivingNorwayHTML(fileLocation, fileEncoding = "", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileLocation}}{A \code{character} scalar containing the location of the
HTML file}

\item{\code{fileEncoding}}{A \code{character} scalar. If non-empty, declares the encoding to be used on a file so the
character data can be re-encoded as they are written}

\item{\code{...}}{A set of arguments to be passed to the \code{\link[xml2::xml_new_root]{xml_new_root}}
function to define namespace information.  The values returned by the helper function
\code{\link[getDefaultEMLDefinitionInfo]{getDefaultEMLDefinitionInfo}} will be used unless
overwritten by user input here.  In addition, arguments required by the EML standard but
not provided here (such as the \link[https://ediorg.github.io/data-package-best-practices/EMLmetadata/Root-element.html]{packageID}
attribute) will be otherwise auto-generated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-importFromLivingNorwayRMD"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-importFromLivingNorwayRMD}{}}}
\subsection{Method \code{importFromLivingNorwayRMD()}}{
Retrieve metadata information from a RMD file that has been created with a
Living Norway HTML tag schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$importFromLivingNorwayRMD(fileLocation, fileEncoding = "", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileLocation}}{A \code{character} scalar containing the location of the
RMD file}

\item{\code{fileEncoding}}{A character string. If non-empty, declares the encoding to be used on a file so the
character data can be re-encoded as they are written}

\item{\code{...}}{A set of arguments to be passed to the \code{\link[xml2::xml_new_root]{xml_new_root}}
function to define namespace information.  The values returned by the helper function
\code{\link[getDefaultEMLDefinitionInfo]{getDefaultEMLDefinitionInfo}} will be used unless
overwritten by user input here.  In addition, arguments required by the EML standard but
not provided here (such as the \link[https://ediorg.github.io/data-package-best-practices/EMLmetadata/Root-element.html]{packageID}
attribute) will be otherwise auto-generated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-importFromEML"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-importFromEML}{}}}
\subsection{Method \code{importFromEML()}}{
Retrieve metadata information from an Ecological Metadata Language (EML) file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$importFromEML(fileLocation, fileEncoding = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileLocation}}{A \code{character} scalar containing the location of the
RMD file}

\item{\code{fileEncoding}}{A character string. If non-empty, declares the encoding to be used on a file so the
character data can be re-encoded as they are written}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-importFromDwCArchive"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-importFromDwCArchive}{}}}
\subsection{Method \code{importFromDwCArchive()}}{
Retrieve metadata information from a Darwin Core archive file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$importFromDwCArchive(fileLocation, fileEncoding = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileLocation}}{A \code{character} scalar containing the location of the
RMD file}

\item{\code{fileEncoding}}{A character string. If non-empty, declares the encoding to be used on a file so the
character data can be re-encoded as they are written}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-exportToEML"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-exportToEML}{}}}
\subsection{Method \code{exportToEML()}}{
Export the metadata as an EML XML file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$exportToEML(fileLocation, fileEncoding = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileLocation}}{A \code{character} scalar containing the location to store the EML fiøe}

\item{\code{fileEncoding}}{A character string. If non-empty, declares the encoding to be used on a file so the
character data can be re-encoded as they are written}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-new}{}}}
\subsection{Method \code{new()}}{
Initialise a metadata object from an import file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$new(fileLocation, fileEncoding = "", fileType = NA, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileLocation}}{A \code{character} scalar containing the location of the import file}

\item{\code{fileEncoding}}{A \code{character} string. If non-empty, declares the encoding to be used on a file so the
character data can be re-encoded as they are written}

\item{\code{fileType}}{A \code{character} scalar stating the type of the file.  File type can be either
\code{"rmarkdown"}, \code{"html"}, \code{"eml"}, or \code{"darwincore"}. If \code{NA} then the
file type will be determined from the file extension}

\item{\code{...}}{A set of arguments to be passed to the \code{\link[xml2::xml_new_root]{xml_new_root}}
function to define namespace information.  The values returned by the helper function
\code{\link[getDefaultEMLDefinitionInfo]{getDefaultEMLDefinitionInfo}} will be used unless
overwritten by user input here.  In addition, arguments required by the EML standard but
not provided here (such as the \link[https://ediorg.github.io/data-package-best-practices/EMLmetadata/Root-element.html]{packageID}
attribute) will be otherwise auto-generated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-print"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-print}{}}}
\subsection{Method \code{print()}}{
Print the metadata information to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-getTitle"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-getTitle}{}}}
\subsection{Method \code{getTitle()}}{
Retrieve the title of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$getTitle(lang = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{A \code{character} scalar that specifies the language of the title to return.  This
is useful when the title has multiple translations in the metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-getCreatorInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-getCreatorInfo}{}}}
\subsection{Method \code{getCreatorInfo()}}{
Retrieve the information of the dataset creators
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$getCreatorInfo(lang = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{A \code{character} scalar that specifies the language of the elements to return.  This
is useful when the elements have multiple translations in the metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-getAbstract"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-getAbstract}{}}}
\subsection{Method \code{getAbstract()}}{
Retrieve the dataset abstract
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$getAbstract(lang = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{A \code{character} scalar that specifies the language of the elements to return.  This
is useful when the elements have multiple translations in the metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DwCMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DwCMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
