% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwCMetadata.R
\name{DwCMetadata}
\alias{DwCMetadata}
\title{read_metadata
read the metadata from a rmarkdown file}
\description{
read_metadata
read the metadata from a rmarkdown file

read_metadata
read the metadata from a rmarkdown file
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DwCMetadata$new()}}
\item \href{#method-getMetadata}{\code{DwCMetadata$getMetadata()}}
\item \href{#method-clone}{\code{DwCMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$new(metadata)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-getMetadata}{}}}
\subsection{Method \code{getMetadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$getMetadata(filepath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filepath}}{a filepath to a RMarkdown file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Output: text string of yaml information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DwCMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
