% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_metadata.R
\name{get_data_metadata}
\alias{get_data_metadata}
\title{Functions for mapping to Darwin Core
function for adding data level metadata to the (e.g. institution, country, community, etc)
BORROW FROM EML??}
\usage{
get_data_metadata()
}
\value{
data level metadata in eml
}
\description{
Functions for mapping to Darwin Core
function for adding data level metadata to the (e.g. institution, country, community, etc)
BORROW FROM EML??
}
