% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geographic_extent.R
\name{get_geographic_extent}
\alias{get_geographic_extent}
\title{Get the geographic extent of a dataframe with geographic coordinates}
\usage{
get_geographic_extent(lat, lon)
}
\arguments{
\item{lat}{latitude column indicated by data$lat}

\item{lon}{longitude column indicated by data$lon}
}
\value{
A ggplot map of the coordinates
}
\description{
This function plots a map of the latitude and longitude from a datafile specified by the user.
}
