% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa_cover.R
\name{get_taxa_cover}
\alias{get_taxa_cover}
\title{get_taxa_cover
Get taxonomic coverage from raw data / mapped data.}
\usage{
get_taxa_cover(x)
}
\arguments{
\item{x}{taxonomic column from user defined dataframe}
}
\value{
Output: list of taxa. Potential chart displaying proportion of occurences from different taxa.
}
\description{
get_taxa_cover
Get taxonomic coverage from raw data / mapped data.
}
