% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFdnaDerivedData.R
\name{getGBIFdnaDerivedDataTerm}
\alias{getGBIFdnaDerivedDataTerm}
\title{DNA Derived data Term Information}
\usage{
getGBIFdnaDerivedDataTerm()
}
\value{
A \code{DwCTerm} object containing the term information
}
\description{
Return the information of the term associated with the GBIF dnaDerivedData class
}
\seealso{
\code{\link[DwCTerm]{DwCTerm}}
}
