% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNTagFunctions.R
\name{getTagGenerationFunctions}
\alias{getTagGenerationFunctions}
\title{Retrieve the tag generation functions
Get a full list of functions that generate tags formatted so that metadata scraper can
retrieve and collate the metadata}
\usage{
getTagGenerationFunctions()
}
\value{
\code{List} of functions for each of the available Living Norway metadata tags
}
\description{
Retrieve the tag generation functions
Get a full list of functions that generate tags formatted so that metadata scraper can
retrieve and collate the metadata
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}; Matthew Grainger, \email{matthew.grainger@nina.no}
}
\keyword{internal}
