% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNTagFunctions.R
\name{LNaddTag}
\alias{LNaddTag}
\title{Generate a Living Norway metadata tag
Add a metadata HTML tag.  This function is designed to be run in RMarkdown documents.  This function
is not designed to be run directly by the users but serves as a utility functions for other
HTML tag generation functions.}
\usage{
LNaddTag(
  tagText,
  tagType,
  tagID = NA,
  parentID = NA,
  isHidden = FALSE,
  lang = NA,
  ...
)
}
\arguments{
\item{tagID}{A \code{character} scalar containing the unique ID of the tag}

\item{parentID}{A \code{character} scalar containing the unique ID of the parent of the tag
(if there is a parent)}

\item{isHidden}{A \code{logical} scalar determining whether the tag should appear in the rendered HTML text}

\item{lang}{A \code{character} scalar defining the language attribute to set for the tag. See
\url{https://www.w3.org/International/articles/language-tags/}{the W3C guide to language tags} for the
best format to use for this attribute.  A value \code{NA} indicates that no langauge attribute will be set.}

\item{tagTxt}{A \code{character} scalar containing the text to be enclosed in the tag}

\item{tagtype}{A \code{character} scalar containing the name of the EML element that will be tagged}
}
\value{
A \code{character} scalar containing the rendered output including any HTML tags if a
HTML document is being knitted
}
\description{
Generate a Living Norway metadata tag
Add a metadata HTML tag.  This function is designed to be run in RMarkdown documents.  This function
is not designed to be run directly by the users but serves as a utility functions for other
HTML tag generation functions.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}; Matthew Grainger, \email{matthew.grainger@nina.no}
}
\keyword{internal}
